Move transmute_ptr_to_ptr to its own module
